+++ lib/ripple/application.ex
@@ -13,6 +13,7 @@ defmodule Ripple.Application do
       supervisor(RippleWeb.Endpoint, []),
       # Start station echo
       worker(RippleWeb.Broadcasters.StationBroadcaster, [], restart: :permanent),
+      worker(Ripple.Stations.StationEventListener, [], restart: :permanent),
       # Start your own worker by calling: Ripple.Worker.start_link(arg1, arg2, arg3)
       # worker(Ripple.Worker, [arg1, arg2, arg3]),
       {Horde.Registry, name: Ripple.StationRegistry},