+++ lib/ripple_web/views/track_view.ex
@@ -38,4 +38,17 @@ defmodule RippleWeb.TrackView do
       }
     }
   end
+
+  def render("track_history.json", %{track: track}) do
+    %{
+      artwork_url: track.artwork_url,
+      duration: track.duration,
+      dj: render_one(track.dj, UserView, "user.json"),
+      started_at: track.started_at,
+      finished_at: track.finished_at,
+      name: track.name,
+      provider: track.provider,
+      url: track.url
+    }
+  end
 end